feat: Remove burning and handling fee
